Also rename tests.
