+++ airbyte-workers/src/test/java/io/airbyte/workers/internal/StateDeltaTrackerTest.java
@@ -2,9 +2,8 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.protocols.airbyte;
+package io.airbyte.workers.internal;
 
-import io.airbyte.workers.internal.StateDeltaTracker;
 import io.airbyte.workers.internal.StateDeltaTracker.StateDeltaTrackerException;
 import java.util.Collections;
 import java.util.HashMap;
